Add FallbackToDisk option
* When a static resource cannot be found FallbackToDisk when set
to true will allow falling back to look for the resource on disk.
This is particularily useful when you have mixes assets of both static
and on disk resources i.e. js, css etc are static but you have a dir
for user avatar images which is on disk and needs to be read and written
to.